Move temporary tests to permanent location for dataPermission4
