+++ src/Databrary/Model/Permission/Types.hs
@@ -114,13 +114,6 @@ instance Monoid Access where
 
 deriveLiftMany [''Permission, ''Access]
 
-{-
-data VolumeAccessPolicy = PublicRestricted | SharedRestricted | PermLevelDefault
-  deriving (Show, Eq)
-
-deriveLift ''VolumeAccessPolicy
--}
-
 data PublicPolicy = PublicRestrictedPolicy | PublicNoPolicy deriving (Show, Eq)
 
 data SharedPolicy = SharedRestrictedPolicy | SharedNoPolicy deriving (Show, Eq)
